move the class `Backoff` to the project Kinesis connector
